Deleting this migrations file.
The migrations file is in the .gitignore.  I am deleting this here, because my team members and I think that pushing/pulling this file to and from this repo is making Github flag our accounts.  We will have to run the makemigrations, and migrate commands whenever we pull down to test code or merge an updated master.